Updated Generate, Added pdf file with info, Added bash script
Modified the generate.py file to automatically created the .js files in a state where they can be instantly run.

Added the pdf file that the python file reads from to generate the .js files.

Added a script file that when run, uses generate.py to automatically generate the .js files & runs them.